requirements.txt
langchain 
flask 
langchain-community 
langchain-groq 
faiss-cpu
langchain-huggingface
python-dotenv
beautifulsoup4